solutions up to 701
Another update to search that allowed me to get solutions up to 701.  As before, the search uses a "frontier" approach that allows it to suspend a search branch and resume it when the search has reduced the target to a potentially achievable level.  The search frontier is stored on disk if it gets too big.  But now we significantly reduce the size of the search space by looking ahead to weed out individual candidates that aren't compatible with the numbers already selected.

I've tried to optimize the search code to eliminate synchronization that was preventing the code from making optimal use of a large number of cores.

Also made significant improvements to memoization to set a configurable limit on the size of the cache.